requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=4.5=1_gnu
astor=0.8.1=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.7.5=h06a4308_1
certifi=2021.5.30=py39h06a4308_0
charset-normalizer=2.0.4=pypi_0
click=8.0.1=pypi_0
colorama=0.4.4=pypi_0
colorlog=5.0.1=pypi_0
cudatoolkit=11.1.74=h6bb024c_0
decorator=5.0.9=pypi_0
dill=0.3.4=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.0.12=pypi_0
freetype=2.10.4=h5ab3b9f_0
gast=0.4.0=pypi_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
h5py=3.3.0=pypi_0
huggingface-hub=0.0.12=pypi_0
idna=3.2=pypi_0
intel-openmp=2021.3.0=h06a4308_3350
jieba=0.42.1=pypi_0
joblib=1.0.1=pypi_0
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libffi=3.3=he6710b0_2
libgcc-ng=9.3.0=h5101ec6_17
libgomp=9.3.0=h5101ec6_17
libiconv=1.15=h63c8f33_5
libidn2=2.3.2=h7f8727e_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.3.0=hd4cf53a_17
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h295c915_1
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.0=py39h42c9631_2
mkl_random=1.2.2=py39h51133e4_0
multiprocess=0.70.12.2=pypi_0
ncurses=6.2=he6710b0_1
nettle=3.7.3=hbbd107a_1
ninja=1.10.2=hff7bd54_1
nltk=3.6.2=pypi_0
numpy=1.20.3=py39hf144106_0
numpy-base=1.20.3=py39h74d4b33_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openjpeg=2.3.0=h05c96fa_1
openssl=1.1.1k=h27cfd23_0
packaging=21.0=pypi_0
paddlenlp=2.0.7=pypi_0
paddlepaddle-gpu=2.1.2=pypi_0
pillow=8.3.1=py39h2c7a002_0
pip=21.2.3=py36h06a4308_0
protobuf=3.17.3=pypi_0
pyparsing=2.4.7=pypi_0
python=3.9.6=h12debd9_0
pytorch=1.9.0=py3.9_cuda11.1_cudnn8.0.5_0
pyyaml=5.4.1=pypi_0
readline=8.1=h27cfd23_0
regex=2021.8.3=pypi_0
requests=2.26.0=pypi_0
sacremoses=0.0.45=pypi_0
scikit-learn=0.24.2=pypi_0
scipy=1.7.1=pypi_0
seqeval=1.2.2=pypi_0
setuptools=52.0.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.36.0=hc218d9a_0
threadpoolctl=2.2.0=pypi_0
tk=8.6.10=hbc83047_0
tokenizers=0.10.3=pypi_0
torchaudio=0.9.0=py39
torchvision=0.10.0=py39_cu111
tqdm=4.62.1=pypi_0
transformers=4.9.2=pypi_0
typing_extensions=3.10.0.0=pyh06a4308_0
tzdata=2021a=h5d7bf9c_0
urllib3=1.26.6=pypi_0
wheel=0.37.0=pyhd3eb1b0_0
xz=5.2.5=h7b6447c_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
